+++ front/context/src/AccountsContext.tsx
@@ -2,19 +2,32 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { web3Accounts, web3Enable } from '@polkadot/extension-dapp';
+import { DerivedStakingAccount } from '@polkadot/api-derive/types';
 import {
   InjectedAccountWithMeta,
   InjectedExtension,
 } from '@polkadot/extension-inject/types';
-import React, { createContext, useState } from 'react';
+import { logger } from '@polkadot/util';
+import { decodeAddress, encodeAddress } from '@polkadot/util-crypto';
+import React, { createContext, useContext, useEffect, useState } from 'react';
+import { distinctUntilChanged, take } from 'rxjs/operators';
+
+import { ApiContext } from './ApiContext';
+import { SystemContext } from './SystemContext';
+import { IS_SSR } from './util';
+
+export interface DecoratedAccount
+  extends InjectedAccountWithMeta,
+    DerivedStakingAccount {}
 
 interface AccountsContext {
-  accounts: InjectedAccountWithMeta[];
+  decoratedAccounts: DecoratedAccount[];
   readonly extension: InjectedExtension;
   fetchAccounts: () => Promise<void>;
   isExtensionReady: boolean;
 }
+
+const l = logger('accounts-context');
 
 export const AccountsContext = createContext({} as AccountsContext);
 
@@ -29,36 +42,82 @@ interface Props {
 
 export function AccountsContextProvider(props: Props): React.ReactElement {
   const { children, originName } = props;
+  const { api, isApiReady } = useContext(ApiContext);
+  const { chain } = useContext(SystemContext);
   const [accounts, setAccounts] = useState<InjectedAccountWithMeta[]>([]);
+  const [decoratedAccounts, setDecoratedAccounts] = useState<
+    DecoratedAccount[]
+  >([]);
   const [extension, setExtension] = useState<InjectedExtension>();
   const [isReady, setIsReady] = useState(false);
+
+  useEffect(() => {
+    if (!isApiReady || !accounts) {
+      return;
+    } else {
+      // make sure it's encoded correctly
+      accounts.map((account: InjectedAccountWithMeta) => {
+        account.address = encodeAddress(decodeAddress(account.address), 2);
+
+        const sub = api.derive.staking
+          .account(account.address)
+          .pipe(take(1), distinctUntilChanged())
+          .subscribe((derivedStakingAccount: DerivedStakingAccount) => {
+            setDecoratedAccounts([
+              ...decoratedAccounts,
+              {
+                ...account,
+                ...derivedStakingAccount,
+              },
+            ]);
+          });
+
+        return () => sub.unsubscribe();
+      });
+    }
+  }, [accounts, isApiReady]);
 
   /**
    * Fetch accounts from the extension
    */
   async function fetchAccounts(): Promise<void> {
-    const extensions = await web3Enable(originName);
-
-    if (!extensions.length) {
-      throw new Error(
-        'No extension found. Please install PolkadotJS extension.'
+    if (typeof window !== `undefined`) {
+      const { web3Accounts, web3Enable } = await import(
+        '@polkadot/extension-dapp'
       );
-    }
 
-    setExtension(extensions[0]);
-    setIsReady(true);
-    setAccounts(await web3Accounts());
+      const extensions = await web3Enable(originName);
+
+      if (!extensions.length) {
+        throw new Error(
+          'No extension found. Please install PolkadotJS extension.'
+        );
+      }
+
+      if (IS_SSR) {
+        throw new Error('Window does not exist during SSR');
+      }
+
+      setExtension(extensions[0]);
+      setAccounts(await web3Accounts());
+      l.log(`Accounts ready, encoded to ss58 prefix of ${chain}`);
+      setIsReady(true);
+    }
   }
 
   return (
     <AccountsContext.Provider
       value={{
-        accounts,
+        decoratedAccounts,
         get extension(): InjectedExtension {
           if (!extension) {
             throw new Error(
               'Please use `extension` only after `isExtensionReady` is set to true'
             );
+          }
+
+          if (IS_SSR) {
+            throw new Error('Window does not exist during SSR');
           }
 
           return extension;