+++ front/gatsby/gatsby-browser.js
@@ -2,6 +2,6 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-export * from './provider';
-export * from './ssr';
-export * from './ui';
+import 'semantic-ui-css/semantic.min.css';
+
+export { wrapPageElement, wrapRootElement } from './gatsby-wrappers';