+++ front/gatsby/gatsby-ssr.js
@@ -2,6 +2,4 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-export * from './provider';
-export * from './ssr';
-export * from './ui';
+export { wrapPageElement, wrapRootElement } from './gatsby-wrappers';