+++ front/gatsby/src/components/Layout/Layout.tsx
@@ -2,25 +2,22 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import './Layout.module.css';
-
-import { Grid } from '@substrate/ui-components';
+import { RouteComponentProps } from '@reach/router';
 import React from 'react';
 
 import { Header } from './Header';
-import { Validators } from './Validators';
 
-interface Props {
+interface Props extends RouteComponentProps {
   children: React.ReactNode;
 }
 
-export function Layout({ children }: Props): React.ReactElement {
+export function Layout(props: Props): React.ReactElement {
+  const { children } = props;
+
   return (
-    <Grid container>
-      <Header />
-      {/* TODO use react router */}
-      <Validators />
+    <>
+      <Header {...props} />
       {children}
-    </Grid>
+    </>
   );
 }